test(ui): decompose broad component harnesses
