+++ ui/package.json
@@ -85,6 +85,10 @@
     "@radix-ui/react-tooltip": "^1.2.8",
     "@tanstack/react-query": "^5.90.2",
     "@xyflow/react": "^12.10.2",
+    "@you-agent-factory/client": "file:./packages/client",
+    "@you-agent-factory/factory-emulator": "file:./packages/factory-emulator",
+    "@you-agent-factory/factory-replay": "file:./packages/factory-replay",
+    "@you-agent-factory/factory-visualizers": "file:./packages/factory-visualizers",
     "cron-parser": "4.9.0",
     "d3": "^7.9.0",
     "elkjs": "^0.11.1",
