+++ ui/packages/factory-replay/src/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@you-agent-factory/client": ["../../client/src/index.ts"]
+    }
+  }
+}
