+++ ui/src/features/factory-emulator/components/customer-factory-emulator-demos-lifecycle.test.tsx
@@ -7,8 +7,9 @@ import {
   waitFor,
   within,
 } from "@testing-library/react";
-import { afterEach, describe, expect, it, vi } from "vitest";
+import { afterEach, describe, expect, it } from "vitest";
 
+import { bunVi as vi } from "../../../testing/bun/vi-compat";
 import { customerFactoryEmulatorDemoFixtures } from "../lib/customer-demo-fixtures";
 import { CustomerFactoryEmulatorDemos } from "./customer-factory-emulator-demos";
