+++ ui/src/features/factory-emulator/components/customer-factory-emulator-demos.test.tsx
@@ -8,8 +8,9 @@ import {
 } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import type { FactoryEmulatorScenario } from "@you-agent-factory/factory-emulator";
-import { afterEach, describe, expect, it, vi } from "vitest";
+import { afterEach, describe, expect, it } from "vitest";
 
+import { bunVi as vi } from "../../../testing/bun/vi-compat";
 import { customerFactoryEmulatorDemoFixtures } from "../lib/customer-demo-fixtures";
 import { CustomerFactoryEmulatorDemos } from "./customer-factory-emulator-demos";
