+++ ui/src/testing/bun/vi-compat.ts
@@ -1,6 +1,28 @@
-import { mock, spyOn, type Mock } from "bun:test";
+import { type Mock, mock, spyOn } from "bun:test";
 
 const stubbedGlobals = new Map<PropertyKey, PropertyDescriptor | undefined>();
+const stubbedWindowGlobals = new Map<
+  PropertyKey,
+  PropertyDescriptor | undefined
+>();
+
+function secondaryWindowGlobal(): object | null {
+  const candidate = (globalThis as typeof globalThis & { window?: object })
+    .window;
+  return candidate && candidate !== globalThis ? candidate : null;
+}
+
+function restoreDescriptor(
+  target: object,
+  key: PropertyKey,
+  descriptor: PropertyDescriptor | undefined,
+) {
+  if (descriptor) {
+    Object.defineProperty(target, key, descriptor);
+  } else {
+    Reflect.deleteProperty(target, key);
+  }
+}
 
 export const bunVi = {
   clearAllMocks: mock.clearAllMocks,
@@ -12,25 +34,39 @@ export const bunVi = {
   restoreAllMocks: mock.restore,
   stubGlobal(key: PropertyKey, value: unknown) {
     if (!stubbedGlobals.has(key)) {
-      stubbedGlobals.set(
-        key,
-        Object.getOwnPropertyDescriptor(globalThis, key),
-      );
+      stubbedGlobals.set(key, Object.getOwnPropertyDescriptor(globalThis, key));
     }
     Object.defineProperty(globalThis, key, {
       configurable: true,
       writable: true,
       value,
     });
+    const testWindow = secondaryWindowGlobal();
+    if (testWindow) {
+      if (!stubbedWindowGlobals.has(key)) {
+        stubbedWindowGlobals.set(
+          key,
+          Object.getOwnPropertyDescriptor(testWindow, key),
+        );
+      }
+      Object.defineProperty(testWindow, key, {
+        configurable: true,
+        writable: true,
+        value,
+      });
+    }
   },
   unstubAllGlobals() {
     for (const [key, descriptor] of stubbedGlobals) {
-      if (descriptor) {
-        Object.defineProperty(globalThis, key, descriptor);
-      } else {
-        Reflect.deleteProperty(globalThis, key);
+      restoreDescriptor(globalThis, key, descriptor);
+    }
+    const testWindow = secondaryWindowGlobal();
+    if (testWindow) {
+      for (const [key, descriptor] of stubbedWindowGlobals) {
+        restoreDescriptor(testWindow, key, descriptor);
       }
     }
     stubbedGlobals.clear();
+    stubbedWindowGlobals.clear();
   },
 };
